Add usage steps and file purpose comments
