+++ ioc_collector/__init__.py
@@ -1,0 +1,2 @@
+"""Pacote que contém a aplicação IOC Collector."""
+
