+++ ioc_collector/collectors/__init__.py
@@ -1,0 +1,2 @@
+"""Subpacote responsável pelos coletores de dados externos."""
+
