+++ ioc_collector/requirements.txt
@@ -1,2 +1,3 @@
 python-dotenv
 requests
+rich
