+++ ioc_collector/utils/__init__.py
@@ -1,0 +1,2 @@
+"""Funções auxiliares e utilidades do IOC Collector."""
+
